Arreglando decimales del Porciento y las ganancias del usuario

+++ Class/MyPlan.cs
@@ -2,7 +2,7 @@ namespace Meta_xi.Application;
 public class MyPlan
 {
     public required string Name { get ; set ; }
-    public required float Percentage { get ; set ; }
+    public required double Percentage { get ; set ; }
     public required int DaysRemaining { get ; set ; }
     public required double HourBenefit { get ; set ; }
     public required double DailyBenefit { get ; set ; }
